fix: Sprint 11B production hardening — dashboard filter, payment sync, schema, logging cleanup
- Fix 6 dashboard queries: .neq('payment_status','failed') → .eq('payment_status','completed')
- failPaymentSale now also updates payment_transactions.status to 'failed' for consistency
- Remove .int() from STK Push amount validation (Tuma accepts decimals)
- Remove investigation-only response body dumps and phone format diagnostic logs
- Keep correlationId, structured errors, and payment lifecycle logs

+++ app/api/payments/tuma/stk-push/route.ts
@@ -26,7 +26,7 @@ import { z } from 'zod'
 const stkPushSchema = z.object({
   saleId: z.string().uuid(),
   phoneNumber: z.string().regex(/^(254|0|\+254)\d{9}$/, 'Phone must be a valid Safaricom number (e.g. 0712345678)'),
-  amount: z.number().int().positive().max(150000),
+  amount: z.number().positive().max(150000),
   description: z.string().max(20).optional(),
 })
